Link DLL in x64 Mode
Now that we're using:
  ./Launch-VsDevShell.ps1 -Arch amd64 -HostArch amd64
We have the DLL linking natively in 64-bit-Windows mode.
(This means it can link with the 64-bit vcpkg libraries)

Started using em_socket_t/evutil_socket_t properly (for a socket fd).

Created a standalone em_socket_t.h (previously was inside
eventmeth.h).

Fixed external symbol problem in pist_timelog.h (for __unDName).

Added needed Windows libraries to meson.build.

Fixed up pist_resource.cc and pist_ifaddrs.cc so they actually compile
the code they contain (previously was ifdef-ed out by mistake).

Now need to start using the proper functions for open, close, read and
write for Windows sockets.

+++ include/pistache/transport.h
@@ -166,7 +166,7 @@ namespace Pistache::Tcp
                 return _raw;
             }
 
-            BufferHolder detach(size_t offset = 0)
+            BufferHolder detach(off_t offset = 0)
             {
                 if (!isRaw())
                     return BufferHolder(_fd, size_, offset);
